requirements.txt
Flask==2.0.2
livereload==2.5.1
flask-mongoengine==1.0.0
python-dotenv==0.19.2
pymongo==3.12.3
bcrypt==3.2.0
